* dev: (626 commits)
  AO-138 added some android:ids to reach some layouts and views in app-view
  [AN-886] Request to latest version
  [UNDO-f97532339ea288ed895ba23ee87e45af6cb5dec2] reverting last commit that was causing a new bug in the Downloads tab.
  version name was increased
  Trello - Drawer - Login
  [AN-883] Fixed a bug: notifications stay no longer when cancel is pressed.
  Fixed GetAdsRequest.java bug.
  created new cache with file saving and loading mechanism + unit tests for a few minor cases. further unit and integration tests are needed
  [AN-884] AppViewIntallWidget fix: now sends Download.COMPLETED events & knock
  Added Partner Id as argument to GetAdsRequest.java.
  [AN-860] Removed some flurry events (too much spam)
  Trello - Onboarding - Third Step - or Login (podemos afastar mais o link do botão e do call to action?
  [AN-879] Appstimline Update card is no longer listening for the downloads state so it doesn't show errors when an error as occurred when the user didn't update trought the timeline.
  AO-145 private to protected on button listeners, Needed to override color of buttons
  unit tests not running due to strange missing dependency
  Added excluded package to appview suggested ads.
  AO-119 Added relative layout id for removing mature content switch on Partners
  [AN-815 871] - Removing progressbar from read all reviews/comments screen. Incoming response to listreviews request no longer sends screen to first results.
  [AN-874] Installation wizard is now launch correctly if download recovers from an error during the process
  minor improvements to hit the cache
  ...

# Conflicts:
#	v8engine/build.gradle
#	v8engine/src/main/AndroidManifest.xml
#	v8engine/src/main/java/cm/aptoide/pt/v8engine/fragment/implementations/AppViewFragment.java
#	v8engine/src/main/res/menu/menu_appview_fragment.xml
#	v8engine/src/main/res/values/strings.xml

+++ downloadmanager/src/main/java/cm/aptoide/pt/downloadmanager/CacheHelper.java
@@ -1,36 +1,60 @@
+/*
+ * Copyright (c) 2016.
+ * Modified by SithEngineer on 02/09/2016.
+ */
+
 package cm.aptoide.pt.downloadmanager;
 
-import java.io.File;
-
-import cm.aptoide.pt.database.Database;
+import android.text.format.DateUtils;
+import cm.aptoide.pt.database.accessors.DownloadAccessor;
 import cm.aptoide.pt.database.realm.Download;
 import cm.aptoide.pt.database.realm.FileToDownload;
+import cm.aptoide.pt.downloadmanager.interfaces.CacheManager;
 import cm.aptoide.pt.downloadmanager.interfaces.DownloadSettingsInterface;
+import cm.aptoide.pt.logger.Logger;
 import cm.aptoide.pt.utils.FileUtils;
-import io.realm.Realm;
-import io.realm.RealmResults;
 import io.realm.Sort;
-import lombok.Cleanup;
+import java.io.File;
+import lombok.AllArgsConstructor;
 
 /**
  * Created by trinkes on 7/7/16.
  */
-public class CacheHelper {
+@AllArgsConstructor public class CacheHelper implements CacheManager {
 
-	public static void cleanCache(DownloadSettingsInterface settingsInterface, String cacheDirPath) {
-		long maxCacheSize = settingsInterface.getMaxCacheSize() * 1024 * 1024 * 8;
-		@Cleanup
-		Realm realm = Database.get();
-		RealmResults<Download> allSorted = realm.where(Download.class).findAllSorted("timeStamp", Sort.ASCENDING);
+  private static final int VALUE_TO_CONVERT_MB_TO_BYTES = 1024 * 1024;
+  public static String TAG = CacheHelper.class.getSimpleName();
+  private DownloadAccessor downloadAccessor;
+  private DownloadSettingsInterface dirSettings;
 
-		int i = 0;
-		while (i < allSorted.size() - 1 && FileUtils.dirSize(new File(cacheDirPath)) > maxCacheSize) {
-			Download download = allSorted.get(i);
-			for (final FileToDownload fileToDownload : download.getFilesToDownload()) {
-				FileUtils.removeFile(fileToDownload.getFilePath());
-			}
-			Database.delete(download, realm);
-			i++;
-		}
-	}
+  public void cleanCache() {
+    long maxCacheSize = dirSettings.getMaxCacheSize() * VALUE_TO_CONVERT_MB_TO_BYTES;
+    String cacheDirPath = dirSettings.getDownloadDir();
+    long now = System.currentTimeMillis();
+
+    downloadAccessor.getAllSorted(Sort.ASCENDING).first().map(downloads -> {
+      int i = 0;
+      while (i < downloads.size() - 1
+          && FileUtils.dirSize(new File(cacheDirPath)) > maxCacheSize
+          && (now - downloads.get(i).getTimeStamp()) > DateUtils.HOUR_IN_MILLIS) {
+
+        Download download = downloads.get(i);
+        for (final FileToDownload fileToDownload : download.getFilesToDownload()) {
+          if (!(fileToDownload.getFileType() == FileToDownload.OBB
+              && download.getOverallDownloadStatus() == Download.COMPLETED)) {
+            FileUtils.removeFile(fileToDownload.getFilePath());
+          }
+        }
+        downloadAccessor.delete(download.getMd5());
+        i++;
+      }
+      return i;
+    }).subscribe(numberDeletedFiles -> {
+      if (numberDeletedFiles > 0) {
+        Logger.d(TAG, "Cache cleaned: " + numberDeletedFiles);
+      } else {
+        Logger.d(TAG, "Cache not cleaned");
+      }
+    }, throwable -> throwable.printStackTrace());
+  }
 }
